Feat: add stage block
